Changed fonts and added background color.
A pretty hefty revision to the whole design philosophy of the website -
Going for a sharper look with a larger contrast in color gives a
stronger impression in my mind.

Removing the giant transparent piano overlay also did much to get the
site not-looking like something from the 2008 anymore!

To do: Revise bio, make sure everything is 3 paragraphs long to keep the
page balance.